pyproject.toml
[tool.poetry]
name = "mech23"
version = "0.1.0"
description = ""
authors = ["na-trium-144 <100704180+na-trium-144@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.16.4"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
